+++ src-tauri/src/models/mod.rs
@@ -1,2 +1,3 @@
+pub mod installed_program;
 pub mod system_info;
 pub mod virtualization;